The best test with the paddle
This is using the ball as a paddle since we lose missiles with paddles on the standard kernel. This felt great but the aiming stopping point needs to be vertically inverted so the stopping point of the ball is between south south west and south. Right now when you aim with the paddle the stopping point is between NNW and N. The left most turn of the paddle should be SSW and the right most run of the paddle should be W.

Fix this before merging with main Nebulords branch.